pyproject.toml
[tool.poetry]
name = "hasura-workshop"
version = "1.0"
description = ""
authors = ["Your Name <diego.cousin@xmartlabs.com>"]

[tool.poetry.dependencies]
python = "3.10.2"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
PyJWT = "^2.6.0"

